tests/wpt/tests/css/css-position/sticky/position-sticky-input-box-gets-focused-after-scroll.html

<!DOCTYPE html>
<title>Focusing on visible sticky input box should scroll to unshifted sticky position.</title>
<link rel="help" href="https://www.w3.org/TR/css-position-3/#sticky-pos" />
<meta name="assert" content="This test checks that focusing on visible sticky
positioned input box should not scroll the page." />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
body {
  height: 2000px;
}

input {
  position: sticky;
  top: 10px;
}
</style>

<input type="text" id="input"/>

<script>
test(() => {
  var input = document.getElementById('input');
  window.scrollTo(0, 100);
  input.focus();
  assert_equals(window.scrollY, 0);
}, 'Focusing on visible sticky input box should reset the scroll to unshifted sticky position.');
</script>
