LayoutTests/imported/w3c/web-platform-tests/css/css-position/sticky/position-sticky-input-box-gets-focused-after-scroll-expected.txt



FAIL Focusing on visible sticky input box should reset the scroll to unshifted sticky position. assert_equals: expected 0 but got 100
